+++ CHANGELOG.md
@@ -47,7 +47,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Added `IoU` class interface ([#4704](https://github.com/PyTorchLightning/pytorch-lightning/pull/4704))
 
 
-- Added `predictions` can be returned in `trainer.test(...)` ([#5468](https://github.com/PyTorchLightning/pytorch-lightning/pull/5468))
+- Added mechanism to return `predictions` in `results` object from `trainer.test(...)` using `self.add_predictions` in `LightningModule` ([#5468](https://github.com/PyTorchLightning/pytorch-lightning/pull/5468))
 
 
 ### Changed